+++ src/tn/esprit/java/PO/AutoveicoloPO.java
@@ -3,7 +3,7 @@ package tn.esprit.java.PO;
 public class AutoveicoloPO extends VeicoloPO {
     public int Nbr_door;
 
-    public AutoveicoloPO(int nbr_telaio, String marca, String modello, long iva, int nbr_door) {
+    public AutoveicoloPO(int nbr_telaio, String marca, String modello, int iva, int nbr_door) {
         super (nbr_telaio, marca, modello, iva);
         Nbr_door = nbr_door;
     }
@@ -15,7 +15,7 @@ public class AutoveicoloPO extends VeicoloPO {
         Nbr_door = nbr_door;
     }
 
-    public AutoveicoloPO( String marca, String modello, long iva, int nbr_door) {
+    public AutoveicoloPO( String marca, String modello, int iva, int nbr_door) {
         super( marca, modello, iva);
         Nbr_door = nbr_door;
     }
